# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ちょっとレンタルサービスのAPI仕様書",
        "title": "Chotto Rental API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/applications/{application_id}/status": {
            "put": {
                "description": "応募のステータスを更新します（案件作成者のみ可能）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "応募"
                ],
                "summary": "応募ステータス更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "応募ID",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新するステータス情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_application.UpdateApplicationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Application"
                        }
                    },
                    "400": {
                        "description": "無効なパラメータ",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "403": {
                        "description": "この応募のステータスを更新する権限がありません\",",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "応募が見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "メールアドレスとパスワードで認証し、トークンを発行します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認証"
                ],
                "summary": "ログイン",
                "parameters": [
                    {
                        "description": "ログイン情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ログイン成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handler_auth.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "無効なパラメータ",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "メールアドレスまたはパスワードが間違っています",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "新規ユーザーを登録し、認証トークンを発行します",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "認証"
                ],
                "summary": "ユーザー登録",
                "parameters": [
                    {
                        "description": "登録情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "登録成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handler_auth.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "無効なパラメータ",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "409": {
                        "description": "メールアドレスが既に使用されています",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
        "/cases/{id}": {
            "put": {
                "description": "指定された案件の情報を更新します（案件作成者のみ可能）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "案件"
                ],
                "summary": "案件更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer {token} 形式",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "案件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新する案件情報",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_case.CreateCaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Case"
                        }
                    },
                    "400": {
                        "description": "リクエストが不正です",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "401": {
                        "description": "認証エラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "403": {
                        "description": "この操作を行う権限がありません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "404": {
                        "description": "案件が見つかりません",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    },
                    "500": {
                        "description": "サーバーエラー",
                        "schema": {
                            "$ref": "#/definitions/github_com_kageyama0_chotto-rental_pkg_util.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_kageyama0_chotto-rental_internal_model.Application": {
            "description": "応募情報",
            "type": "object",
            "properties": {
                "applicant": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "applicantID": {
                    "type": "string"
                },
                "case": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.Case"
                },
                "caseID": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_internal_model.Case": {
            "description": "案件情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "durationMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "reward": {
                    "type": "integer"
                },
                "scheduledDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_kageyama0_chotto-rental_internal_model.User"
                },
                "userID": {
                    "description": "案件の依頼者",
                    "type": "string"
                }
            }
        },
...
        "github_com_kageyama0_chotto-rental_internal_model.User": {
            "description": "ユーザー情報",
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "displayName": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "noShowCount": {
                    "type": "integer"
                },
                "passwordHash": {
                    "type": "string"
                },
                "trustScore": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_kageyama0_chotto-rental_pkg_util.Response": {
            "description": "HTTPレスポンス",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
        "internal_handler_application.UpdateApplicationStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "accepted",
                        "rejected"
                    ]
                }
            }
        },
        "internal_handler_auth.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {}
            }
        },
        "internal_handler_auth.LoginRequest": {
            "description": "ログインリクエスト",
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_handler_auth.RegisterRequest": {
            "description": "ユーザー登録リクエスト",
            "type": "object",
            "required": [
                "display_name",
                "email",
                "password"
            ],
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "internal_handler_case.CreateCaseRequest": {
            "type": "object",
            "required": [
                "description",
                "duration_minutes",
                "location",
                "reward",
                "scheduled_date",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer",
                    "minimum": 1
                },
                "location": {
                    "type": "string"
                },
                "reward": {
                    "type": "integer",
                    "minimum": 0
                },
                "scheduled_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "「Bearer 」の後にJWTトークンを付与してください",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}